베타 0.1
- ReadMe 파일 업데이트
- 알림 패널 제작
- 접속 보상 알림 생성
- Coin 자료형 (int -> ulong)
- SpawnPrice 계산 공식 변경 (SpawnLevel 추가)
- https://mergeslime.netlify.app

+++ MergeSlime/Assets/Scripts/Slime/Mining.cs
@@ -33,7 +33,7 @@ public class Mining : MonoBehaviour
         yield return new WaitForSeconds(maxMiningT);
 
         SetMining();
-        slime.dataManager.coin.GainCoin(miningAmount);
+        slime.dataManager.coin.GainCoin((ulong)miningAmount);
         slime.spawnManager.SpawnMoneyTxt(slime.body.transform, miningAmount);
 
         StartCoroutine(MiningRoutine());
